Moved GlowEvents to Events
