+++ ExtraEnemyCustomization/Events/Inject/Inject_EnemyAppearance_InterpolateGlows.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Text;
 using UnityEngine;
 
-namespace EECustom.Customizations.Models.Inject
+namespace EECustom.Events.Inject
 {
     [HarmonyPatch(typeof(EnemyAppearance))]
     class Inject_EnemyAppearance_InterpolateGlows
@@ -14,7 +14,7 @@ namespace EECustom.Customizations.Models.Inject
         [HarmonyWrapSafe]
         static void Pre_InterpolateGlow(ref Color col, Vector4 pos, float transitionTime, EnemyAppearance __instance)
         {
-            col = GlowEvents.FireEvent(__instance.m_owner, col, pos);
+            col = EnemyGlowEvents.FireEvent(__instance.m_owner, col, pos);
         }
     }
 }
